requirements.txt
python-dotenv
PyPDF2
scikit-learn
pyresparser
pdfminer3
requests
langchain_community